requirements.txt
# Core
python-dotenv>=1.0.0

# LLM API
openai>=1.6.0

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# Vector database
chromadb>=0.4.22

# Text processing
tiktoken>=0.5.2

# Utilities
numpy>=1.24.0
